platform/core: add support for IPv4 PtP/peer addresses
OpenVPN's 'p2p' topology does send a different peer address than the
local address for tunX, plus the server may also push routes that
use the peer address as the next hop.  Unless NetworkManager configures
the tunX interface with the peer address, the kernel refuses to add
the routes with the peer nexthop, because the kernel has no idea
how to talk to the peer, because the peers address is not reachable
over any routes (none have yet been added) and the peer's address is
not assigned to any interface yet.

OR: could we just add a host route through the tunX pointing to
the peer address, and leave the PtP peer address unset?