libnm-util: if one of slave-type and master is set, both must be set
slave-type is required because master may refer to an interface that
is not yet created, and thus the details of the slave connection are
ambiguous.  While we can auto-detect slave-type in some cases (like
if a bridge-port setting exists, then slave-type=bridge), that
functionality isn't implemented yet and doesn't work for all cases.
So for the moment, require that both slave-type and master are
set if one is set.