core: wait to manage !IFF_UP software devices until they come up (rh #1030947)
NM sets a device IFF_UP when it manages the device, to ensure things like link
detection work and that firmware is loaded for the device.  Unfortunately, this
prevents adding bridge ports or bond slaves outside NetworkManager if the
bridge/bond was created outside NM.  Since we'd really like to cooperate better
with external changes, don't begin to manage the interface until it's set IFF_UP
(and thus actually usable) by whatever created it.