wifi: fix restarted AP not being found again (bgo #733105)
If the supplicant removes an AP, WPAS_REMOVED_TAG is set for the
AP and it becomes available for NM to delete when it reaches the
max age.  But if the AP is found again before being deleted, the
tag is not reset and the AP could be deleted prematurely.

https://bugzilla.gnome.org/show_bug.cgi?id=733105