platform: define the largest hardware address the platform supports
NM_UTILS_HWADDR_LEN_MAX is really the largest address size that
nm_utils_hwaddr_aton() supports, which isn't really appropriate for
the platform to use since it may do different things with hardware
addresses.  But since if one changes we should evaluate whether to
change the other one, ensure they are the same with a static assert.