platform: remove nm_platform_get_type_name() function
This function mostly returned a name based on the link-type.
In those cases, where it didn't, it was also not clearly defined
what string we would expect and how to use it.

If we later have a specific need to get a descriptive string for
an unknown type, we should add a new function especially for
the use-case at hand.