platform: remove type_name field from NMPlatformLink
Before, @type could be UNKNOWN and @type_name could be a more
descriptive type obtained from the rtnl_link structure.
Just drop that, if the type enum is UNKNOWN, we don't need an alternate,
opaque type description.

Especially, because of the uglyness that in such cases, @type_name
would not be a static string, but an allocated string from the rtnl_link
structure.