platform: don't return more specific link-type from link_extract_type() when unknown
link_extract_type() would return the rtnl_type if the link-type
could not be detected. One problem is that this string is not
a static string, which can be unexpected.

But also, this "more specific description of type" is not used much
beyond logging. The only place is the @type_description of NMDeviceGeneric.
I think, it's better to add "unknown" types explicitly and one-by-one, like
we do for example for "openvswitch" type. If a type is so uncommon that
we don't even add detection to link_extract_type(), "unknown" is good
enough as a description.

This changes behavior, because it can return the @type_description of
some NMDeviceGeneric to "unknown".