squash! platform: lookup ifname using libc function instead of platform cache
Rework this patch and reword commit message. Current master doesn't
actually have a problem (but platform-refactoring would).

New commit message:
>>>>>

platform: detect TUN/TAP device in link_extract_type() independently of platform cache

link_extract_type() would call tun_get_properties() to determine whether
the link if a TAP or TUN device. The previous implementation would
receive the ifindex, and resolve the ifname via lookup in the platform
cache.
This means, the call on link_extract_type() will only succeed to detect
the TUN/TAP properties, if the libnl object is already in the cache.
Currently that is always the case and there is no problem.

It is desireable, that we can resolve the link type of an object without
consulting the platform cache first.