linux-platform: bump device route metric if another device's route would clash
The kernel always adds a route with metric=0 for an address assigned to a
device. This is not too cool as it will hijack traffic from more preferred
devices. We add a route with a device specific metric to cope with this.
It causes the other route to disappear.

However, we did this only this once. If another device which wanted to use the
same metric cames up, we didn't add another route, since a route with the same
metric would cause the first one to be removed.

This is bad, since the metric=0 route would stay. Even worse, when the first
and the second address are in the same subnet, it would become the preferred
route to that subnet. A rp_filter set to strict (default for Fedora and RHEL)
would cause the traffic that's still being sent to the first interface to be
dropped.

This made Cockpit developers sad.