device: let NM_DEVICE_MASTER be equal to nm_device_get_master()
There is only one user of the NM_DEVICE_MASTER property (NMActiveConnection):
device_master_changed() only uses the property for change notification,
but returns early if !nm_device_get_master().

We might emit the notify::master signal too often, even when nothing
changed (because currently we always emit it when priv->master changes).
Even if we would fix that to only emit the signal when
nm_device_get_master() changes, that would cause not problem because
the only user only cares if the master is set. We could remove spurious
signals if the master is unset without change.

Hence, none of the user actually care about the difference. But change
it because it's ugly that they are not equal in any case.