fixup! vpn-manager: support multiple VPN connections of the same type
I don't think it's a good idea to use the connection UUID. I would
instead use a per-service counter.

Indeed, currently only one UUID is active at a time. But after
applied-connection branch, we could add a D-Bus call to "re-associate" a
NMActiveConnection with another NMSettingsConnection -- without taking
the connection doesn (or touching the applied connection).
The problem thatn isn't that the UUID of the active connection changes,
but that you might re-activate the now-free settings-connection -- and
get a conflict.