WIP: fixup! vpn-manager: support multiple VPN connections of the same type
WIP: I just started but I think it should be like the following:

NMVpnConnection should get it's plugin_info at construct-time and it
should not change. It's wrong to only pass it later during
nm_vpn_manager_activate_connection().
That requires some refcatoring...

The "service-type" entry of a NMSettingVpn connection is untrusted.
Use instead the nm_vpn_plugin_info_get_service(), which is a bit more
trusted (as it comes from the .name file).
Actually, they should be identical anyway because that is how we match
the connection to a vpn-plugin.